LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-patternMismatch-expected.txt


PASS [INPUT in TEXT status] The pattern attribute is not set 
PASS [INPUT in TEXT status] The value attibute is empty string 
PASS [INPUT in TEXT status] The value attribute matches the pattern attribute 
PASS [INPUT in TEXT status] The value(ABC) in unicode attribute matches the pattern attribute 
FAIL [INPUT in TEXT status] The value attribute mismatches the pattern attribute assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in TEXT status] The value attribute mismatches the pattern attribute even when a subset matches assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in SEARCH status] The pattern attribute is not set 
PASS [INPUT in SEARCH status] The value attibute is empty string 
PASS [INPUT in SEARCH status] The value attribute matches the pattern attribute 
PASS [INPUT in SEARCH status] The value(ABC) in unicode attribute matches the pattern attribute 
FAIL [INPUT in SEARCH status] The value attribute mismatches the pattern attribute assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in SEARCH status] The value attribute mismatches the pattern attribute even when a subset matches assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in TEL status] The pattern attribute is not set 
PASS [INPUT in TEL status] The value attibute is empty string 
PASS [INPUT in TEL status] The value attribute matches the pattern attribute 
PASS [INPUT in TEL status] The value(ABC) in unicode attribute matches the pattern attribute 
FAIL [INPUT in TEL status] The value attribute mismatches the pattern attribute assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in TEL status] The value attribute mismatches the pattern attribute even when a subset matches assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in URL status] The pattern attribute is not set 
PASS [INPUT in URL status] The value attibute is empty string 
PASS [INPUT in URL status] The value attribute matches the pattern attribute 
PASS [INPUT in URL status] The value(ABC) in unicode attribute matches the pattern attribute 
FAIL [INPUT in URL status] The value attribute mismatches the pattern attribute assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in URL status] The value attribute mismatches the pattern attribute even when a subset matches assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in EMAIL status] The pattern attribute is not set 
PASS [INPUT in EMAIL status] The value attibute is empty string 
PASS [INPUT in EMAIL status] The value attribute matches the pattern attribute 
PASS [INPUT in EMAIL status] The value(ABC) in unicode attribute matches the pattern attribute 
FAIL [INPUT in EMAIL status] The value attribute mismatches the pattern attribute assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in EMAIL status] The value attribute mismatches the pattern attribute even when a subset matches assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in PASSWORD status] The pattern attribute is not set 
PASS [INPUT in PASSWORD status] The value attibute is empty string 
PASS [INPUT in PASSWORD status] The value attribute matches the pattern attribute 
PASS [INPUT in PASSWORD status] The value(ABC) in unicode attribute matches the pattern attribute 
FAIL [INPUT in PASSWORD status] The value attribute mismatches the pattern attribute assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in PASSWORD status] The value attribute mismatches the pattern attribute even when a subset matches assert_true: The validity.patternMismatch should be true, when control is disabled. expected true got false
